internships

+++ src/app/api/internships/[id]/route.ts
@@ -3,7 +3,7 @@ import { type NextRequest } from "next/server";
 import { auth } from "@/auth";
 import prisma from "@/utils/db";
 import { Role } from "@/types/auth";
-import { InternshipWithCompanyLocationSetUser } from "@/types/entities";
+import { InternshipFullRecord, InternshipWithCompanyLocationSetUser } from "@/types/entities";
 
 export async function GET(
   request: NextRequest,
@@ -17,24 +17,76 @@ export async function GET(
     });
   }
 
-  let internship: InternshipWithCompanyLocationSetUser | null =
-    await prisma.internship.findFirst({
+  let internship: InternshipWithCompanyLocationSetUser | InternshipFullRecord | null;
+  if (session.user.role === Role.ADMIN || session.user.role === Role.TEACHER) {
+    internship = await prisma.internship.findFirst({
       include: {
-        user: true,
-        location: true,
         set: true,
+        user: true,
         company: true,
+        location: true,
         reservationUser: true,
       },
       where: { id: id },
     });
+  } else {
+  internship = await prisma.internship.findFirst({
+      select: {
+        id: true,
+        classname: true,
+        created: true,
+        kind: true,
+        userId: true,
+        companyId: true,
+        setId: true,
+        user: {
+          select: {
+            givenName: true,
+            surname: true,
+          },
+        },
+        company: {
+          select: {
+            name: true,
+            companyIdentificationNumber: true,
+          },
+        },
+        location: {
+          select: {
+            municipality: true,
+          },
+        },
+        set: {
+          select: {
+            name: true,
+            year: true,
+            editable: true,
+            active: true,
+            daysTotal: true,
+            hoursDaily: true,
+            start: true,
+            end: true,
+            continuous: true,
+          },
+        },
+      },
+      where: { id: id },
+    });
+  }
   if (!internship) {
     return new Response("Not found", {
       status: 404,
     });
   }
-  // omezit osobní data mimo adminy a konkrétního studenta
-
+  if (
+    session.user.role !== Role.ADMIN &&
+    session.user.role !== Role.TEACHER &&
+    session.user.id !== internship?.userId
+  ) {
+    return new Response("Forbidden", {
+      status: 403,
+    });
+  }
   return Response.json(internship);
 }
 
@@ -50,15 +102,49 @@ export async function DELETE(
       status: 401,
     });
   }
-  let internship: Internship | null = await prisma.internship.findFirst({
-    where: { id: id },
-  });
-
-  if (!internship) {
-    return new Response("Not found", {
-      status: 404,
+  let internship: InternshipWithCompanyLocationSetUser | null =
+    await prisma.internship.findFirst({
+      select: {
+        id: true,
+        classname: true,
+        created: true,
+        kind: true,
+        userId: true,
+        companyId: true,
+        setId: true,
+        user: {
+          select: {
+            givenName: true,
+            surname: true,
+          },
+        },
+        company: {
+          select: {
+            name: true,
+            companyIdentificationNumber: true,
+          },
+        },
+        location: {
+          select: {
+            municipality: true,
+          },
+        },
+        set: {
+          select: {
+            name: true,
+            year: true,
+            editable: true,
+            active: true,
+            daysTotal: true,
+            hoursDaily: true,
+            start: true,
+            end: true,
+            continuous: true,
+          },
+        },
+      },
+      where: { id: id },
     });
-  }
 
   if (
     session.user.role !== Role.ADMIN &&
@@ -68,8 +154,16 @@ export async function DELETE(
       status: 403,
     });
   }
-
-  // TODO omezit možnost smazat data, pokud není sada editovatelná
+if (internship === null) {
+    return new Response("Not found", {
+      status: 404,
+    });
+  }
+  if (internship.set.editable === false) {
+    return new Response("Set of this internship is not editable.", {
+      status: 402,
+    })
+  }
 
   await prisma.internship.delete({
     where: { id: id },
@@ -91,11 +185,7 @@ export async function PUT(
       status: 401,
     });
   }
-  if (session.user.role !== "admin") {
-    return new Response("Forbidden", {
-      status: 403,
-    });
-  }
+  
   let internship: InternshipWithCompanyLocationSetUser | null =
     await prisma.internship.findFirst({
       select: {
@@ -144,10 +234,7 @@ export async function PUT(
       status: 404,
     });
   }
-  if (
-    session.user.role !== Role.ADMIN &&
-    session.user.id !== internship?.userId
-  ) {
+  if (session.user.role !== Role.ADMIN && session.user.role !== Role.TEACHER && session.user.id !== internship?.userId) {
     return new Response("Forbidden", {
       status: 403,
     });
@@ -160,6 +247,14 @@ export async function PUT(
   }
   const body = await request.json();
   body.kind = Number(body.kind);
+
+  if (session.user.role !== Role.ADMIN && session.user.role !== Role.TEACHER) {
+    body.reservationUserId = undefined;
+    body.highlighted = undefined;
+  }
+
+  body.updated = new Date();
+
   await prisma.internship.update({
     where: { id: id },
     data: body,
